backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Azure SDK
azure-identity==1.15.0
azure-mgmt-compute==30.4.0
azure-mgmt-network==25.2.0
azure-mgmt-storage==21.1.0
azure-mgmt-resource==23.0.1
azure-mgmt-containerservice==28.0.0
azure-mgmt-rdbms==10.2.0
azure-mgmt-sql==4.0.0b13
azure-mgmt-cosmosdb==9.3.0

# Utilities
pydantic==2.5.3
httpx==0.26.0
aiohttp==3.9.1

# Cryptography for SSH key generation
cryptography==41.0.7

# CORS
starlette==0.35.1

